+++ extensions/viewer3d/src/OHIFDicomM3DViewport.js
@@ -4,6 +4,7 @@ import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 
 import dicomParser from 'dicom-parser';
+import dcmjs from 'dcmjs';
 
 import OHIF from '@ohif/core';
 import { eventTypes as uiEvents } from '@ohif/ui';
@@ -11,7 +12,7 @@ import { str2ab } from '@ohif/core';
 import { LoadingIndicator } from '@ohif/extension-vtk';
 
 import M3DModelView from './threejs/M3DModelView.js';
-import { SOP_CLASS_UIDS } from './OHIFDicom3DSopClassHandler.js';
+import { SOP_CLASS_UIDS, MIMETYPE_STL } from './OHIFDicom3DSopClassHandler.js';
 
 import './styles/LoadingIndicator.css';
 
@@ -19,7 +20,8 @@ const { DicomLoaderService, createEncapsulatedDocumentFileUrl } = OHIF.utils;
 
 class OHIFDicomM3DViewport extends Component {
   // OHIF viewport with support for displaying 3D models and scenes. Supports STL and GLB.
-  // TODO: Add support for USDZ files.
+  // STL files may be part of a series with multiple models.
+  // TODO: Implement support for USDZ files.
 
   static propTypes = {
     studies: PropTypes.object,
@@ -34,9 +36,9 @@ class OHIFDicomM3DViewport extends Component {
 
   state = {
     byteArray: null,
-    modelDataset: null,
-    modelFileUrl: null,
-    modelType: null,
+    models: [],
+    modelType: '',
+    modelCount: 0,
     error: null,
     isLoaded: false,
     cinePlaying: false,
@@ -60,6 +62,7 @@ class OHIFDicomM3DViewport extends Component {
     this.setState({ isLoaded: true });
 
     if (
+      this.api.sceneData &&
       this.api.sceneData.animations &&
       this.api.sceneData.animations.length &&
       _.isFunction(this.props.setViewportSpecificData)
@@ -119,20 +122,45 @@ class OHIFDicomM3DViewport extends Component {
 
   getOrCreate3DModel(byteArray) {
     // Parse the provided byte array to a dataset and file URL
+    const { displaySet } = this.props.viewportData;
+    const { series } = displaySet;
+
     const modelDataset = this.parseByteArray(byteArray);
     const modelType = modelDataset
       ? modelDataset.string('x00420012')
       : undefined;
+
+    // Parse instance properties to retrieve display properties
+    const SOPInstanceUID = modelDataset
+      ? modelDataset.string('x00080018')
+      : undefined;
+    let modelColor;
+    if (SOPInstanceUID) {
+      const instance = series.getInstanceByUID(SOPInstanceUID);
+      const idata = instance.getData().metadata;
+      modelColor = idata.RecommendedDisplayCIELabValue
+        ? OHIF.utils.color.rgb2hex(
+            ...dcmjs.data.Colors.dicomlab2RGB(
+              idata.RecommendedDisplayCIELabValue
+            ).map((x) => Math.round(x * 255))
+          )
+        : undefined;
+    }
+
+    // const colorRgb = modelDataset ? modelDataset.number('x0062000d') : undefined;
+    // Create blob URL
     const modelFileUrl = this.get3DModelFileUrl(modelDataset, byteArray, {
       mimetype: modelType,
     });
 
-    return { byteArray, modelDataset, modelType, modelFileUrl };
+    return { byteArray, modelDataset, modelType, modelFileUrl, modelColor };
   }
 
   fetchModel() {
-    // Retrieve models from the DICOM series
-    const { displaySet, studies } = this.props.viewportData;
+    // Retrieve DICOM model instances.
+
+    const { displaySet, studies, metadata } = this.props.viewportData;
+    const { numImageFrames, series } = displaySet;
 
     // File available from cache, retrieve and set inline byte array
     if (displaySet.metadata && displaySet.metadata.EncapsulatedDocument) {
@@ -142,21 +170,92 @@ class OHIFDicomM3DViewport extends Component {
       if (InlineBinary) {
         // Create dataset from DICOM inline binary data
         const inlineBinaryData = atob(InlineBinary);
-        this.setState({ ...this.getOrCreate3DModel(str2ab(inlineBinaryData)) });
+        const mdata = this.getOrCreate3DModel(str2ab(inlineBinaryData));
+        this.setState({
+          modelCount: 1,
+          models: [mdata],
+          modelType: mdata.modelType,
+          percentComplete: 100,
+        });
         return;
       }
     }
 
-    // Retrieve GLB files from remote server
-    DicomLoaderService.findDicomDataPromise(displaySet, studies).then(
-      (data) => {
-        this.setState({ ...this.getOrCreate3DModel(new Uint8Array(data)) });
-      },
-      (error) => {
-        this.setState({ error });
-        throw new Error(error);
-      }
-    );
+    if (!numImageFrames && _.isUndefined(numImageFrames)) {
+      this.setState({ modelCount: 1 });
+
+      // Retrieve single model file from remote server
+      DicomLoaderService.findDicomDataPromise(displaySet, studies).then(
+        (data) => {
+          const mdata = this.getOrCreate3DModel(new Uint8Array(data));
+          this.setState({
+            models: [mdata],
+            percentComplete: 100,
+            modelType: mdata.modelType,
+          });
+        },
+        (error) => {
+          this.setState({ error });
+          throw new Error(error);
+        }
+      );
+    } else if (numImageFrames && numImageFrames > 1) {
+      this.setState({ modelCount: numImageFrames });
+
+      // Retrieve multiple models from remote server
+      const _component = this;
+
+      _.times(numImageFrames, function (i) {
+        // Retrieve DICOM instance
+        const instance = series.getInstanceByIndex(i);
+
+        // Create a copy of the display set and add instance specific data.
+        const displayInstance = _.extend(_.clone(displaySet), {
+          wadoUri: instance.getData().wadouri,
+          SOPInstanceUID: instance.getSOPInstanceUID(),
+        });
+
+        // Retrieve DICOM instances for all models
+        DicomLoaderService.findDicomDataPromise(displayInstance, studies).then(
+          (data) => {
+            const { percentComplete, modelCount, modelType, models } =
+              _component.state;
+
+            // Retrieve model instance from DICOM data
+            const mdata = _component.getOrCreate3DModel(new Uint8Array(data));
+
+            // Create update to component state
+            let newState = {
+              models: [...models, mdata],
+              percentComplete:
+                models.length + 1 == numImageFrames
+                  ? 100
+                  : Math.round(percentComplete + 100 * (1 / modelCount)),
+            };
+            if (!modelType) {
+              newState.modelType = mdata.modelType;
+            }
+
+            // Add -90 degree x transform for STL models so that
+            // the major axis (x) will point upwards and the model will appear to be
+            // facing the viewer. If the transform is not applied, the orbit controls
+            // have an extremely limited range.
+            if (modelType == MIMETYPE_STL) {
+              newState.coordinateTransform = {
+                rotation: [-Math.PI / 2, 0, 0],
+              };
+            }
+
+            // Update component state
+            _component.setState(newState);
+          },
+          (error) => {
+            _component.setState({ error });
+            throw new Error(error);
+          }
+        );
+      });
+    }
   }
 
   onInteractionStart() {
@@ -231,8 +330,7 @@ class OHIFDicomM3DViewport extends Component {
       this.setState({
         isLoaded: false,
         byteArray: null,
-        modelDataset: null,
-        modelFileUrl: null,
+        models: [],
         error: null,
         cinePlaying: false,
         cineFrameRate: 60,
@@ -297,17 +395,25 @@ class OHIFDicomM3DViewport extends Component {
   }
 
   render() {
-    const { byteArray, error, models } = this.state;
+    const { byteArray, error, models, percentComplete, modelCount } =
+      this.state;
     const style = { width: '100%', height: '100%', position: 'relative' };
 
     return (
       <>
         <div className="ohif-m3d-model-container" style={style}>
-          {!this.state.isLoaded && <LoadingIndicator />}
-          {this.state.modelFileUrl && (
+          {!this.state.isLoaded && (
+            <LoadingIndicator
+              percentComplete={
+                modelCount && modelCount > 1 ? percentComplete : undefined
+              }
+            />
+          )}
+          {this.state.modelType && this.state.percentComplete == 100 && (
             <M3DModelView
-              modelFileUrl={this.state.modelFileUrl}
               modelType={this.state.modelType}
+              models={this.state.models}
+              coordinateTransform={this.state.coordinateTransform}
               onCreated={this.onModelLoaded.bind(this)}
               onInteractionStart={this.onInteractionStart.bind(this)}
               getStaticUrl={this.props.getStaticUrl}