+++ extensions/dicom-segmentation/CHANGELOG.md
@@ -3,6 +3,29 @@
 All notable changes to this project will be documented in this file. See
 [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.21](https://github.com/OHIF/Viewers/compare/@ohif/extension-dicom-segmentation@0.7.20...@ohif/extension-dicom-segmentation@0.7.21) (2023-01-17)
+
+**Note:** Version bump only for package @ohif/extension-dicom-segmentation
+
+## [0.7.20](https://github.com/OHIF/Viewers/compare/@ohif/extension-dicom-segmentation@0.7.19...@ohif/extension-dicom-segmentation@0.7.20) (2023-01-10)
+
+**Note:** Version bump only for package @ohif/extension-dicom-segmentation
+
+### Bug Fixes
+
+- [#2964](https://github.com/OHIF/Viewers/issues/2964) Update message for
+  segmentation error loading due to orientation tolerance
+  ([#2982](https://github.com/OHIF/Viewers/issues/2982))
+  ([545161d](https://github.com/OHIF/Viewers/commit/545161d5e8d92bb6652f17dc24a36dac8aa7b3a4))
+
+## [0.7.19](https://github.com/OHIF/Viewers/compare/@ohif/extension-dicom-segmentation@0.7.18...@ohif/extension-dicom-segmentation@0.7.19) (2022-12-21)
+
+**Note:** Version bump only for package @ohif/extension-dicom-segmentation
+
+## [0.7.18](https://github.com/OHIF/Viewers/compare/@ohif/extension-dicom-segmentation@0.7.17...@ohif/extension-dicom-segmentation@0.7.18) (2022-12-21)
+
+**Note:** Version bump only for package @ohif/extension-dicom-segmentation
+
 ## [0.7.17](https://github.com/OHIF/Viewers/compare/@ohif/extension-dicom-segmentation@0.7.16...@ohif/extension-dicom-segmentation@0.7.17) (2022-12-12)
 
 ### Bug Fixes